Move GraphParser and NodeParser
