+++ geex-models/src/main/java/nl/tudelft/context/model/graph/parser/NodeParser.java
@@ -1,4 +1,6 @@
-package nl.tudelft.context.model.graph;
+package nl.tudelft.context.model.graph.parser;
+
+import nl.tudelft.context.model.graph.Node;
 
 import java.util.Arrays;
 import java.util.HashSet;
