+++ geex-models/src/test/java/nl/tudelft/context/model/graph/GraphTest.java
@@ -1,5 +1,6 @@
 package nl.tudelft.context.model.graph;
 
+import nl.tudelft.context.model.graph.parser.GraphParser;
 import org.junit.BeforeClass;
 import org.junit.Test;
