+++ geex-models/src/test/java/nl/tudelft/context/model/graph/InsertDeleteGraphTest.java
@@ -1,6 +1,7 @@
 package nl.tudelft.context.model.graph;
 
 import de.saxsys.javafx.test.JfxRunner;
+import nl.tudelft.context.model.graph.parser.GraphParser;
 import nl.tudelft.context.service.LoadService;
 import org.junit.Before;
 import org.junit.Test;
